app: update logging
